Move socket handling into its own folder
