+++ lib/socket/index.js
@@ -1,6 +1,8 @@
 const debug = require('debug')('@feathersjs/transport-commons');
-const channels = require('./channels');
-const routing = require('./routing');
+
+const channels = require('../channels');
+const routing = require('../routing');
+
 const { getDispatcher, runMethod } = require('./utils');
 
 module.exports = function ({ done, emit, socketKey, getParams }) {
