+++ lib/socket/utils.js
